update week 4
